add admin routes to store controller

+++ src/main/java/com/markdoyle/marketplace/repositories/CarRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarRepository extends CrudRepository<Car, Long> {
+public interface CarRepository extends CrudRepository<Car, Integer> {
 	Car findByName(String name);
 }
